# apps/local-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexura Local Server API",
    "version": "0.0.1",
    "description": "The per-branch local server's real API surface -- LAN-only, offline-first (Auth/Distribution doc Part 2: this server has no public URL; it's reachable only on the hotel's own network). Every path below was extracted directly from the real Express route definitions in `src/routes/*.ts` (see `scripts/generate-openapi.mjs`), not hand-typed -- run `npm run openapi:generate` after changing routes to keep this in sync. Auth is a JWT in an httpOnly `access_token` cookie (12h, with an 8h offline-continuation grace period -- see `src/auth/tokens.ts`), not a bearer token in the traditional OpenAPI sense; represented here as a cookie API key for tooling compatibility.",
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local dev server"
    },
    {
      "url": "http://nexura.local:4000",
      "description": "Real branch deployment (Auth doc Part 2, Layer 3) -- LAN hostname, no public DNS/TLS. `localhost` above is genuinely correct for dev, not a placeholder; this second entry is the real production shape."
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
  "paths": {
...
    "/guest-messages/in-house": {
      "get": {
        "operationId": "getGuestMessagesInHouse",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "In-house guest picker for starting a conversation (CO-02)",
        "description": "**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/guest-messages/threads": {
      "get": {
        "operationId": "getGuestMessagesThreads",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "Every guest conversation thread for the branch",
        "description": "Newest activity first, real last-message preview, flags whether the guest is currently in-house.\n\n**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/guest-messages/threads/{guestId}": {
      "get": {
        "operationId": "getGuestMessagesThreadsByGuestId",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "Thread detail + full message history",
        "description": "Lazily creates an empty thread on first access if this guest has never been messaged.\n\n**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "guestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/guest-messages/threads/{guestId}/escalate": {
      "post": {
        "operationId": "postGuestMessagesThreadsByGuestIdEscalate",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "Escalate a thread to management",
        "description": "Posts a real message into a real DM with every Manager/ORG at the branch, reusing Internal Chat rather than a separate inbox.\n\n**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "guestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/guest-messages/threads/{guestId}/forward": {
      "post": {
        "operationId": "postGuestMessagesThreadsByGuestIdForward",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "Forward a thread to a department",
        "description": "Posts a real message into that department's actual Internal Chat channel (auto-created if the branch doesn't have one yet) — the receiving team sees it where they already look.\n\n**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "guestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "department"
                ],
                "properties": {
                  "department": {
                    "type": "string",
                    "enum": [
                      "Front Desk",
                      "Housekeeping",
                      "Maintenance",
                      "Restaurant"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/guest-messages/threads/{guestId}/messages": {
      "post": {
        "operationId": "postGuestMessagesThreadsByGuestIdMessages",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "Log a guest communication",
        "description": "This is a real communication *log*, not a WhatsApp/SMS send — `channel` records how the real-world conversation happened, it doesn't trigger delivery. See the module description.\n\n**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "guestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "channel",
                  "direction",
                  "body"
                ],
                "properties": {
                  "channel": {
                    "type": "string",
                    "enum": [
                      "whatsapp",
                      "sms",
                      "internal"
                    ]
                  },
                  "direction": {
                    "type": "string",
                    "enum": [
                      "to_guest",
                      "from_guest"
                    ]
                  },
                  "body": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/guest-messages/threads/{guestId}/resolve": {
      "post": {
        "operationId": "postGuestMessagesThreadsByGuestIdResolve",
        "tags": [
          "Guest Messaging"
        ],
        "summary": "Mark a guest thread resolved",
        "description": "**Requires permission:** `guests:message`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "guestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
  }
}